backend/requirements.txt
flask==3.0.0
tensorflow-cpu==2.15.0
pillow==10.2.0
numpy==1.26.3
werkzeug==3.0.1